+++ src/main/java/com/mobble/mobbleserver/domain/clubMember/ClubMember.java
@@ -3,7 +3,7 @@ package com.mobble.mobbleserver.domain.clubMember;
 import com.mobble.mobbleserver.common.baseEntity.BaseEntity;
 import com.mobble.mobbleserver.domain.club.core.Club;
 import com.mobble.mobbleserver.domain.member.Member;
-import com.mobble.mobbleserver.refactor.article.entity.ArticleType;
+import com.mobble.mobbleserver.domain.article.ArticleType;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
