+++ src/main/java/com/mobble/mobbleserver/infrastructure/persistence/article/ArticleQueryDslRepositoryImpl.java
@@ -15,12 +15,12 @@ import java.util.Map;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-import static com.mobble.mobbleserver.refactor.article.entity.QArticle.article;
+import static com.mobble.mobbleserver.domain.article.QArticle.article;
 import static com.mobble.mobbleserver.refactor.like.articleLike.entity.QArticleLike.articleLike;
 
 @Repository
 @RequiredArgsConstructor
-public class ArticleRepositoryImpl implements ArticleQueryRepository {
+public class ArticleQueryDslRepositoryImpl implements ArticleQueryDslRepository {
 
     private final JPAQueryFactory queryFactory;
