+++ src/main/java/com/mobble/mobbleserver/infrastructure/web/article/dto/response/ArticleResponseDto.java
@@ -1,9 +1,9 @@
-package com.mobble.mobbleserver.refactor.article.dto.response;
+package com.mobble.mobbleserver.infrastructure.web.article.dto.response;
 
 import com.mobble.mobbleserver.infrastructure.web.comment.dto.response.RootCommentResponseDto;
 import com.mobble.mobbleserver.domain.article.Article;
 import com.mobble.mobbleserver.domain.article.ArticleType;
-import com.mobble.mobbleserver.refactor.article.repository.dto.ArticleLikeInfoDto;
+import com.mobble.mobbleserver.infrastructure.persistence.article.projection.ArticleLikeInfoDto;
 
 import java.time.LocalDateTime;
 import java.util.List;
