+++ src/main/java/com/mobble/mobbleserver/infrastructure/web/comment/dto/request/CommentRequestDto.java
@@ -2,7 +2,7 @@ package com.mobble.mobbleserver.infrastructure.web.comment.dto.request;
 
 import com.mobble.mobbleserver.domain.comment.Comment;
 import com.mobble.mobbleserver.domain.member.Member;
-import com.mobble.mobbleserver.refactor.article.entity.Article;
+import com.mobble.mobbleserver.domain.article.Article;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
